Module Nmea__Parse

val parse : string -> Nmea.Sentence.t

parse s parses the sentence s

val parse_opt : string -> Nmea.Sentence.t option

parse_opt s parses the sentence s, failsafe

val next : Stdlib.in_channel -> Nmea.Sentence.t option

next ch returns the next sentence read from ch

val next_coord : Stdlib.in_channel -> Nmea.Coord.t option

next_coord ch returns the next coordinate read from ch